xong edit and delete

+++ src/main/java/cg/model/Product.java
@@ -9,24 +9,24 @@ import javax.persistence.Id;
 public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)   //tạo id tự động
-    private int id ;
+    private Long id ;
     private String name;
     private float price;
 
     public Product() {
     }
 
-    public Product(int id, String name, float price) {
+    public Product(Long id, String name, float price) {
         this.id = id;
         this.name = name;
         this.price = price;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
